+++ matminer/featurizers/site.py
@@ -2169,126 +2169,6 @@ class LocalPropertyDifference(BaseFeaturizer):
     def implementors(self):
         return ["Logan Ward", "Aik Rui Tan"]
 
-
-class BondOrientationalParameter(BaseFeaturizer):
-    """
-    Averages of spherical harmonics of local neighbors
-
-    Bond Orientational Parameters (BOPs) describe the local environment around an atom by
-    considering the local symmetry of the bonds as computed using spherical harmonics.
-    To create descriptors that are invariant to rotating the coordinate system, we use the
-    average of all spherical harmonics of a certain degree - following the approach of
-    `Steinhardt et al. <https://link.aps.org/doi/10.1103/PhysRevB.28.784>`_.
-    We weigh the contributions of each neighbor with the solid angle of the Voronoi tessellation
-    (see `Mickel et al. <https://aip.scitation.org/doi/abs/10.1063/1.4774084>_` for further
-    discussion). The weighing scheme makes these descriptors vary smoothly with small distortions
-    of a crystal structure.
-
-    In addition to the average spherical harmonics, this class can also compute the :math:`W` and
-    :math:`\hat{W}` parameters proposed by `Steinhardt et al. <https://link.aps.org/doi/10.1103/PhysRevB.28.784>`_.
-
-    Attributes:
-        BOOP Q l=<n> - Average spherical harmonic for a certain degree, n.
-        BOOP W l=<n> - W parameter for a certain degree of spherical harmonic, n.
-        BOOP What l=<n> - :math:`\hat{W}` parameter for a certain degree of spherical harmonic, n.
-
-    References:
-        `Steinhardt et al., _PRB_ (1983) <https://link.aps.org/doi/10.1103/PhysRevB.28.784>`_
-        `Seko et al., _PRB_ (2017) <http://link.aps.org/doi/10.1103/PhysRevB.95.144110>`_
-    """
-
-    def __init__(self, max_l=10, compute_w=False, compute_w_hat=False):
-        """
-        Initialize the featurizer
-
-        Args:
-            max_l (int) - Maximum spherical harmonic to consider
-            compute_w (bool) - Whether to compute Ws as well
-            compute_w_hat (bool) - Whether to compute What
-        """
-        self._nn = VoronoiNN(weight="solid_angle")
-        self.max_l = max_l
-        self.compute_W = compute_w
-        self.compute_What = compute_w_hat
-
-    def featurize(self, strc, idx):
-        # Get the nearest neighbors of the atom
-        nns = get_nearest_neighbors(self._nn, strc, idx)
-
-        # Get the polar and azimuthal angles of each face
-        phi = np.arccos([x["poly_info"]["normal"][-1] for x in nns])
-        theta = np.arctan2(
-            [x["poly_info"]["normal"][1] for x in nns],
-            [x["poly_info"]["normal"][0] for x in nns],
-        )
-
-        # Get the weights for each neighbor
-        weights = np.array([x["weight"] for x in nns])
-        weights /= weights.sum()
-
-        # Compute the spherical harmonics for the desired `l`s
-        Qs = []
-        Ws = []
-        for l in range(1, self.max_l + 1):
-            # Average the spherical harmonic over each neighbor, weighted by solid angle
-            qlm = dict((m, np.dot(weights, sph_harm(m, l, theta, phi))) for m in range(-l, l + 1))
-
-            # Compute the average over all m's
-            Qs.append(np.sqrt(np.pi * 4 / (2 * l + 1) * np.sum(np.abs(list(qlm.values())) ** 2)))
-
-            # Compute the W, if desired
-            if self.compute_W or self.compute_What:
-                w = 0
-                # Loop over all non-zero Wigner 3j coefficients
-                for (m1, m2, m3), wcoeff in get_wigner_coeffs(l):
-                    w += qlm[m1] * qlm[m2] * qlm[m3] * wcoeff
-                Ws.append(w.real)
-
-        # Compute Whats, if desired
-        if self.compute_What:
-            Whats = [
-                w / (q / np.sqrt(np.pi * 4 / (2 * l + 1))) ** 3 if abs(q) > 1.0e-6 else 0.0
-                for l, q, w in zip(range(1, self.max_l + 1), Qs, Ws)
-            ]
-
-        # Compile the results. Always returns Qs, and optionally the W/What
-        if self.compute_W:
-            Qs += Ws
-        if self.compute_What:
-            Qs += Whats
-        return Qs
-
-    def feature_labels(self):
-        q_labels = ["BOOP Q l={}".format(l) for l in range(1, self.max_l + 1)]
-        if self.compute_W:
-            q_labels += ["BOOP W l={}".format(l) for l in range(1, self.max_l + 1)]
-        if self.compute_What:
-            q_labels += ["BOOP What l={}".format(l) for l in range(1, self.max_l + 1)]
-        return q_labels
-
-    def citations(self):
-        return [
-            "@article{Seko2017,"
-            "author = {Seko, Atsuto and Hayashi, Hiroyuki and Nakayama, "
-            "Keita and Takahashi, Akira and Tanaka, Isao},"
-            "doi = {10.1103/PhysRevB.95.144110},"
-            "journal = {Physical Review B}, number = {14}, pages = {144110},"
-            "title = {{Representation of compounds for machine-learning prediction of physical properties}},"
-            "url = {http://link.aps.org/doi/10.1103/PhysRevB.95.144110},"
-            "volume = {95},year = {2017}}",
-            "@article{Steinhardt1983,"
-            "author = {Steinhardt, Paul J. and Nelson, David R. and Ronchetti, Marco},"
-            "doi = {10.1103/PhysRevB.28.784}, journal = {Physical Review B},"
-            "month = {jul}, number = {2}, pages = {784--805},"
-            "title = {{Bond-orientational order in liquids and glasses}},"
-            "url = {https://link.aps.org/doi/10.1103/PhysRevB.28.784}, "
-            "volume = {28}, year = {1983}}",
-        ]
-
-    def implementors(self):
-        return ["Logan Ward", "Aidan Thompson"]
-
-
 class SiteElementalProperty(BaseFeaturizer):
     """
     Elemental properties of atom on a certain site
@@ -2416,338 +2296,3 @@ def _iterate_wigner_3j(l):
                 yield m1, m2, m3
 
 
-class AverageBondLength(BaseFeaturizer):
-    """
-    Determines the average bond length between one specific site
-    and all its nearest neighbors using one of pymatgen's NearNeighbor
-    classes. These nearest neighbor calculators return weights related
-    to the proximity of each neighbor to this site. 'Average bond
-    length' of a site is the weighted average of the distance between
-    site and all its nearest neighbors.
-    """
-
-    def __init__(self, method):
-        """
-        Initialize featurizer
-
-        Args:
-            method (NearNeighbor) - subclass under NearNeighbor used to compute nearest neighbors
-        """
-        self.method = method
-
-    def featurize(self, strc, idx):
-        """
-        Get weighted average bond length of a site and all its nearest
-        neighbors.
-
-        Args:
-            strc (Structure): Pymatgen Structure object
-            idx (int): index of target site in structure object
-
-        Returns:
-            average bond length (list)
-        """
-        # Compute nearest neighbors of the indexed site
-        nns = self.method.get_nn_info(strc, idx)
-        if len(nns) == 0:
-            raise IndexError("Input structure has no bonds.")
-
-        weights = [info["weight"] for info in nns]
-        center_coord = strc[idx].coords
-
-        dists = np.linalg.norm(np.subtract([site["site"].coords for site in nns], center_coord), axis=1)
-
-        return [PropertyStats.mean(dists, weights)]
-
-    def feature_labels(self):
-        return ["Average bond length"]
-
-    def citations(self):
-        return [
-            "@article{jong_chen_notestine_persson_ceder_jain_asta_gamst_2016,"
-            "title={A Statistical Learning Framework for Materials Science: "
-            "Application to Elastic Moduli of k-nary Inorganic Polycrystalline Compounds}, "
-            "volume={6}, DOI={10.1038/srep34256}, number={1}, journal={Scientific Reports}, "
-            "author={Jong, Maarten De and Chen, Wei and Notestine, Randy and Persson, "
-            "Kristin and Ceder, Gerbrand and Jain, Anubhav and Asta, Mark and Gamst, Anthony}, "
-            "year={2016}, month={Mar}}"
-        ]
-
-    def implementors(self):
-        return ["Logan Ward", "Aik Rui Tan"]
-
-
-class AverageBondAngle(BaseFeaturizer):
-    """
-    Determines the average bond angles of a specific site with
-    its nearest neighbors using one of pymatgen's NearNeighbor
-    classes. Neighbors that are adjacent to each other are stored
-    and angle between them are computed. 'Average bond angle' of
-    a site is the mean bond angle between all its nearest neighbors.
-    """
-
-    def __init__(self, method):
-        """
-        Initialize featurizer
-
-        Args:
-            method (NearNeighbor) - subclass under NearNeighbor used to compute nearest
-                                    neighbors
-        """
-        self.method = method
-
-    def featurize(self, strc, idx):
-        """
-        Get average bond length of a site and all its nearest
-        neighbors.
-
-        Args:
-            strc (Structure): Pymatgen Structure object
-            idx (int): index of target site in structure object
-
-        Returns:
-            average bond length (list)
-        """
-        # Compute nearest neighbors of the indexed site
-        nns = self.method.get_nn_info(strc, idx)
-        if len(nns) == 0:
-            raise IndexError("Input structure has no bonds.")
-        center = strc[idx].coords
-
-        sites = [i["site"].coords for i in nns]
-
-        # Calculate bond angles for each neighbor
-        bond_angles = np.empty((len(sites), len(sites)))
-        bond_angles.fill(np.nan)
-        for a, a_site in enumerate(sites):
-            for b, b_site in enumerate(sites):
-                if b == a:
-                    continue
-                dot = np.dot(a_site - center, b_site - center) / (
-                    np.linalg.norm(a_site - center) * np.linalg.norm(b_site - center)
-                )
-                if np.isnan(np.arccos(dot)):
-                    bond_angles[a, b] = bond_angles[b, a] = np.arccos(round(dot, 5))
-                else:
-                    bond_angles[a, b] = bond_angles[b, a] = np.arccos(dot)
-        # Take the minimum bond angle of each neighbor
-        minimum_bond_angles = np.nanmin(bond_angles, axis=1)
-
-        return [PropertyStats.mean(minimum_bond_angles)]
-
-    def feature_labels(self):
-        return ["Average bond angle"]
-
-    def citations(self):
-        return [
-            "@article{jong_chen_notestine_persson_ceder_jain_asta_gamst_2016,"
-            "title={A Statistical Learning Framework for Materials Science: "
-            "Application to Elastic Moduli of k-nary Inorganic Polycrystalline Compounds}, "
-            "volume={6}, DOI={10.1038/srep34256}, number={1}, journal={Scientific Reports}, "
-            "author={Jong, Maarten De and Chen, Wei and Notestine, Randy and Persson, "
-            "Kristin and Ceder, Gerbrand and Jain, Anubhav and Asta, Mark and Gamst, Anthony}, "
-            "year={2016}, month={Mar}}"
-        ]
-
-    def implementors(self):
-        return ["Logan Ward", "Aik Rui Tan"]
-
-
-class SOAP(BaseFeaturizer):
-    """
-    Smooth overlap of atomic positions (interface via DScribe).
-
-    Class for generating a partial power spectrum from Smooth Overlap of Atomic
-    Orbitals (SOAP). This implementation uses real (tesseral) spherical
-    harmonics as the angular basis set and provides two orthonormalized
-    alternatives for the radial basis functions: spherical primitive gaussian
-    type orbitals ("gto") or the polynomial basis set ("polynomial"). By
-    default the faster gto-basis is used. Please see the DScribe SOAP
-    documentation for more details.
-
-    Note that SOAP is only featurized for elements identified by "fit" (see
-    following), thus "fit" must be called before "featurize", or else an error
-    will be raised.
-
-    Based originally on the following publications:
-
-    "On representing chemical environments, Albert P. Bartók, Risi
-        Kondor, and Gábor Csányi, Phys. Rev. B 87, 184115, (2013),
-        https://doi.org/10.1103/PhysRevB.87.184115
-
-    "Comparing molecules and solids across structural and alchemical
-        space", Sandip De, Albert P. Bartók, Gábor Csányi and Michele Ceriotti,
-        Phys.  Chem. Chem. Phys. 18, 13754 (2016),
-        https://doi.org/10.1039/c6cp00415f
-
-    Implementation (and some documentation) originally based on DScribe:
-    https://github.com/SINGROUP/dscribe.
-
-    "DScribe: Library of descriptors for machine learning in materials science",
-        Himanen, L., J{\"a}ger, M. O.J., Morooka, E. V., Federici
-        Canova, F., Ranawat, Y. S., Gao, D. Z., Rinke, P. and Foster, A. S.
-        Computer Physics Communications, 106949 (2019),
-        https://doi.org/10.1016/j.cpc.2019.106949
-
-    Args:
-        rcut (float): A cutoff for local region in angstroms. Should be
-            bigger than 1 angstrom.
-        nmax (int): The number of radial basis functions.
-        lmax (int): The maximum degree of spherical harmonics.
-        sigma (float): The standard deviation of the gaussians used to expand the
-            atomic density.
-        rbf (str): The radial basis functions to use. The available options are:
-
-            * "gto": Spherical gaussian type orbitals defined as :math:`g_{nl}(r) = \sum_{n'=1}^{n_\mathrm{max}}\,\\beta_{nn'l} r^l e^{-\\alpha_{n'l}r^2}`
-            * "polynomial": Polynomial basis defined as :math:`g_{n}(r) = \sum_{n'=1}^{n_\mathrm{max}}\,\\beta_{nn'} (r-r_\mathrm{cut})^{n'+2}`
-
-        periodic (bool): Determines whether the system is considered to be
-            periodic.
-        crossover (bool): Determines if crossover of atomic types should
-            be included in the power spectrum. If enabled, the power
-            spectrum is calculated over all unique species combinations Z
-            and Z'. If disabled, the power spectrum does not contain
-            cross-species information and is only run over each unique
-            species Z. Turned on by default to correspond to the original
-            definition
-    """
-
-    @requires(
-        dscribe,
-        "SOAPFeaturizer requires DScribe. Install from github.com/SINGROUP/dscribe",
-    )
-    def __init__(
-        self,
-        rcut,
-        nmax,
-        lmax,
-        sigma,
-        periodic,
-        rbf="gto",
-        crossover=True,
-    ):
-        self.rcut = rcut
-        self.nmax = nmax
-        self.lmax = lmax
-        self.sigma = sigma
-        self.rbf = rbf
-        self.periodic = periodic
-        self.crossover = crossover
-        self.adaptor = AseAtomsAdaptor()
-        self.length = None
-        self.atomic_numbers = None
-        self.soap = None
-        self.n_elements = None
-
-    @classmethod
-    def from_preset(cls, preset):
-        """
-        Create a SOAP featurizer object from sensible or published presets.
-        Args:
-            preset (str): Choose from:
-                "formation energy": Preset used for formation energy prediction
-                    in the original Dscribe paper.
-        Returns:
-        """
-        valid_presets = ["formation_energy"]
-        if preset == "formation_energy":
-            return cls(6, 8, 8, 0.4, True, "gto", True)
-        else:
-            raise ValueError(f"'{preset}' is not a valid preset. Choose from {valid_presets}")
-
-    def _check_fitted(self):
-        if not self.soap:
-            raise NotFittedError("Please fit SOAP before featurizing.")
-
-    def fit(self, X, y=None):
-        """
-        Fit the SOAP featurizer to a dataframe.
-
-        Args:
-            X ([SiteCollection]): For example, a list of pymatgen Structures.
-            y : unused (added for consistency with overridden method signature)
-
-        Returns:
-            self
-        """
-        # Check that pymatgen.Structures are provided
-        if not all([isinstance(struct, Structure) for struct in X]):
-            raise TypeError("This fit requires an array-like input of Pymatgen " "Structures and sites!")
-
-        elements = set()
-        for s in X:
-            c = s.composition.elements
-            for e in c:
-                if e.Z not in elements:
-                    elements.add(e.Z)
-        self.elements_sorted = sorted(list(elements))
-
-        self.atomic_numbers = elements
-        self.soap = SOAP_dscribe(
-            species=self.atomic_numbers,
-            rcut=self.rcut,
-            nmax=self.nmax,
-            lmax=self.lmax,
-            sigma=self.sigma,
-            rbf=self.rbf,
-            periodic=self.periodic,
-            crossover=self.crossover,
-            average="off",
-            sparse=False,
-        )
-
-        self.length = self.soap.get_number_of_features()
-        return self
-
-    def featurize(self, struct, idx):
-        self._check_fitted()
-        s_ase = self.adaptor.get_atoms(struct)
-        return self.soap.create(s_ase, positions=[idx], n_jobs=self.n_jobs).tolist()[0]
-
-    def feature_labels(self):
-        self._check_fitted()
-        return [f"SOAP_{i}" for i in range(self.length)]
-
-    def citations(self):
-        return [
-            "@article{PhysRevB.87.184115,"
-            "title = {On representing chemical environments},"
-            "author = {Bart'ok, Albert P. and Kondor, Risi and Cs'anyi, "
-            "G'abor},"
-            "journal = {Phys. Rev. B},"
-            "volume = {87},"
-            "issue = {18},"
-            "pages = {184115},"
-            "numpages = {16},"
-            "year = {2013},"
-            "month = {May},"
-            "publisher = {American Physical Society},"
-            "doi = {10.1103/PhysRevB.87.184115},"
-            "url = {https://link.aps.org/doi/10.1103/PhysRevB.87.184115}}",
-            "@Article{C6CP00415F,"
-            "author ={De, Sandip and BartÃ³k, Albert P. and CsÃ¡nyi, GÃ¡bor"
-            " and Ceriotti, Michele},"
-            "title  ={Comparing molecules and solids across structural and "
-            "alchemical space},"
-            "journal = {Phys. Chem. Chem. Phys.},"
-            "year = {2016},"
-            "volume = {18},"
-            "issue = {20},"
-            "pages = {13754-13769},"
-            "publisher = {The Royal Society of Chemistry},"
-            "doi = {10.1039/C6CP00415F},"
-            "url = {http://dx.doi.org/10.1039/C6CP00415F},}",
-            "@article{dscribe, "
-            'author = {Himanen, Lauri and J{"a}ger, Marc O.~J. and '
-            "Morooka, Eiaki V. and Federici Canova, Filippo and Ranawat, "
-            "Yashasvi S. and Gao, David Z. and Rinke, Patrick and Foster, "
-            "Adam S.}, "
-            "title = {{DScribe: Library of descriptors for machine "
-            "learning in materials science}}, "
-            "journal = {Computer Physics Communications}, "
-            "year = {2019}, pages = {106949}, "
-            "doi = {https://doi.org/10.1016/j.cpc.2019.106949}}",
-        ]
-
-    def implementors(self):
-        return ["Lauri Himanen and the DScribe team", "Alex Dunn"]
